+++ minipk.h
@@ -52,16 +52,16 @@ typedef struct {
     size_t size;
 } due_cacheline_t;
 
-typedef int (*user_trap_handler)(trapframe_t*, float_trapframe_t*, long, due_candidates_t*, due_cacheline_t*, word_t*, short, short, short, short, short); //Originally defined in riscv-pk/pk/pk.h
+typedef int (*user_trap_handler)(trapframe_t*, float_trapframe_t*, long, due_candidates_t*, due_cacheline_t*, word_t*, size_t, size_t, int, int, int); //Originally defined in riscv-pk/pk/pk.h
 void dump_tf(trapframe_t* tf); //Originally defined in riscv-pk/pk/pk.h
 int copy_word(word_t* dest, word_t* src); //Originally defined in riscv-pk/pk/pk.h
 int copy_cacheline(due_cacheline_t* dest, due_cacheline_t* src); //Originally defined in riscv-pk/pk/pk.h
 int copy_candidates(due_candidates_t* dest, due_candidates_t* src); //Originally defined in riscv-pk/pk/pk.h
 int copy_trapframe(trapframe_t* dest, trapframe_t* src); //Originally defined in riscv-pk/pk/pk.h
 int copy_float_trapframe(float_trapframe_t* dest, float_trapframe_t* src); //Originally defined in riscv-pk/pk/pk.h
-int load_value_from_message(word_t* recovered_message, word_t* load_value, due_cacheline_t* cl, unsigned load_size, int offset); //Originally defined in riscv-pk/pk/pk.h
-int get_float_register(unsigned frd, unsigned long* raw_value); //Originally defined in riscv-pk/pk/pk.h
-int set_float_register(unsigned frd, unsigned long raw_value); //Originally defined in riscv-pk/pk/pk.h
+int load_value_from_message(word_t* recovered_message, word_t* load_value, due_cacheline_t* cl, size_t load_size, int offset); //Originally defined in riscv-pk/pk/pk.h
+int get_float_register(size_t frd, unsigned long* raw_value); //Originally defined in riscv-pk/pk/pk.h
+int set_float_register(size_t frd, unsigned long raw_value); //Originally defined in riscv-pk/pk/pk.h
 
 extern const char* g_int_regnames[];
 extern const char* g_float_regnames[];
